personal-ai-assistant/requirements.txt
# Web Framework
Flask==3.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0

# Claude Integration  
anthropic==0.8.1
httpx==0.26.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Authentication (Google OAuth)
Flask-Dance==7.0.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0

# Task Scheduling & Background Jobs
celery==5.3.0
redis==5.0.1
schedule==1.2.2

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
python-dateutil==2.8.2

# Development & Testing
pytest==8.0.2
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production
Werkzeug==3.0.1
